requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==1.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Werkzeug==2.3.7
